+++ homeassistant/components/modbus/cover.py
@@ -34,7 +34,7 @@ from .const import (
     CONF_STATUS_REGISTER_TYPE,
     MODBUS_DOMAIN,
 )
-from .modbus_hub import AbstractModbusHub
+from .modbus_base import BaseModbusHub
 
 
 async def async_setup_platform(
@@ -49,7 +49,7 @@ async def async_setup_platform(
 
     covers = []
     for cover in discovery_info[CONF_COVERS]:
-        hub: AbstractModbusHub = hass.data[MODBUS_DOMAIN][discovery_info[CONF_NAME]]
+        hub: BaseModbusHub = hass.data[MODBUS_DOMAIN][discovery_info[CONF_NAME]]
         covers.append(ModbusCover(hub, cover))
 
     async_add_entities(covers)
@@ -60,11 +60,11 @@ class ModbusCover(CoverEntity, RestoreEntity):
 
     def __init__(
         self,
-        hub: AbstractModbusHub,
+        hub: BaseModbusHub,
         config: Dict[str, Any],
     ):
         """Initialize the modbus cover."""
-        self._hub: AbstractModbusHub = hub
+        self._hub: BaseModbusHub = hub
         self._coil = config.get(CALL_TYPE_COIL)
         self._device_class = config.get(CONF_DEVICE_CLASS)
         self._name = config[CONF_NAME]
